Split attribute decode into MetadataPrimitives, keep rendering in Metadata
Step 3 of the shared-primitives migration (docs/metadata-primitives.md). The
DECODE half of attribute reading — resolving an attribute's type name and
decoding its constructor blob to typed argument values, plus the SRM type
providers that drive it — moves to a new AttributeDecoder in the SRM-only
primitives library. The RENDER half — turning those typed values into C#
(`Flags`, `Obsolete("msg")`, `(LayoutKind)0`), the re-emitted-attribute noise
policy, and the scope overloads — stays in Metadata.AttributeReader, which now
calls AttributeDecoder.TryDecode and renders the result.

This is the split that lets a future data consumer (e.g. the analysis layer)
read attribute data without pulling in C# rendering. AttributeReader keeps a
GetAttributeTypeName forwarder so its three other Metadata callers are
unchanged. AttributeDecoder keeps the ILInspector.Metadata namespace, so the
move is reference-only — all builds and tests are green (Metadata 145,
Decompiler 411, dotnet-inspect 1104).
