+++ src/ILInspector.Metadata/AttributeReader.cs
@@ -1,4 +1,3 @@
-using System.Collections.Immutable;
 using System.Reflection;
 using System.Reflection.Metadata;
 using System.Reflection.PortableExecutable;
@@ -163,20 +162,7 @@ public static class AttributeReader
     /// Gets the fully qualified type name of an attribute from its constructor handle.
     /// </summary>
     public static string? GetAttributeTypeName(MetadataReader reader, EntityHandle constructorHandle)
-    {
-        if (constructorHandle.Kind == HandleKind.MemberReference)
-        {
-            var memberRef = reader.GetMemberReference((MemberReferenceHandle)constructorHandle);
-            return TypeResolver.GetTypeName(reader, memberRef.Parent);
-        }
-        else if (constructorHandle.Kind == HandleKind.MethodDefinition)
-        {
-            var methodDef = reader.GetMethodDefinition((MethodDefinitionHandle)constructorHandle);
-            var typeDef = reader.GetTypeDefinition(methodDef.GetDeclaringType());
-            return TypeResolver.GetFullName(reader, typeDef);
-        }
-        return null;
-    }
+        => AttributeDecoder.GetAttributeTypeName(reader, constructorHandle);
 
     /// <summary>
     /// Gets custom attributes for a specific method, identified by type name, method name, and overload index.
@@ -364,29 +350,23 @@ public static class AttributeReader
 
     static string? TryRenderAttribute(MetadataReader reader, CustomAttribute attr)
     {
-        string name = GetShortAttributeName(GetAttributeTypeName(reader, attr.Constructor)!);
-        try
-        {
-            var value = attr.DecodeValue(new AttributeArgTypeProvider(reader));
-            var args = new List<string>();
-            foreach (var arg in value.FixedArguments)
-            {
-                if (RenderArgument(arg.Type, arg.Value) is not { } text)
-                    return null;
-                args.Add(text);
-            }
-            foreach (var named in value.NamedArguments)
-            {
-                if (RenderArgument(named.Type, named.Value) is not { } text)
-                    return null;
-                args.Add($"{named.Name} = {text}");
-            }
-            return args.Count == 0 ? name : $"{name}({string.Join(", ", args)})";
-        }
-        catch
-        {
+        if (AttributeDecoder.TryDecode(reader, attr) is not { } value)
             return null;
+        string name = GetShortAttributeName(GetAttributeTypeName(reader, attr.Constructor)!);
+        var args = new List<string>();
+        foreach (var arg in value.FixedArguments)
+        {
+            if (RenderArgument(arg.Type, arg.Value) is not { } text)
+                return null;
+            args.Add(text);
         }
+        foreach (var named in value.NamedArguments)
+        {
+            if (RenderArgument(named.Type, named.Value) is not { } text)
+                return null;
+            args.Add($"{named.Name} = {text}");
+        }
+        return args.Count == 0 ? name : $"{name}({string.Join(", ", args)})";
     }
 
     /// <summary>Renders one attribute-argument value, or null when its shape is not faithfully spellable (arrays, unknown).</summary>
@@ -431,77 +411,6 @@ public static class AttributeReader
         "System.Reflection.DefaultMemberAttribute" => true,
         _ => false,
     };
-
-    /// <summary>Type provider for attribute-blob decoding: primitives render as C# keywords, everything else as its full name (enums and typeof targets).</summary>
-    sealed class AttributeArgTypeProvider(MetadataReader reader) : ICustomAttributeTypeProvider<string>
-    {
-        public string GetPrimitiveType(PrimitiveTypeCode code) => code switch
-        {
-            PrimitiveTypeCode.Boolean => "bool",
-            PrimitiveTypeCode.Char => "char",
-            PrimitiveTypeCode.SByte => "sbyte",
-            PrimitiveTypeCode.Byte => "byte",
-            PrimitiveTypeCode.Int16 => "short",
-            PrimitiveTypeCode.UInt16 => "ushort",
-            PrimitiveTypeCode.Int32 => "int",
-            PrimitiveTypeCode.UInt32 => "uint",
-            PrimitiveTypeCode.Int64 => "long",
-            PrimitiveTypeCode.UInt64 => "ulong",
-            PrimitiveTypeCode.Single => "float",
-            PrimitiveTypeCode.Double => "double",
-            PrimitiveTypeCode.String => "string",
-            _ => "object",
-        };
-
-        public string GetSystemType() => "System.Type";
-        public bool IsSystemType(string type) => type == "System.Type";
-        public string GetSZArrayType(string elementType) => elementType + "[]";
-        public string GetTypeFromDefinition(MetadataReader r, TypeDefinitionHandle handle, byte rawTypeKind)
-            => TypeResolver.GetFullName(r, r.GetTypeDefinition(handle));
-        public string GetTypeFromReference(MetadataReader r, TypeReferenceHandle handle, byte rawTypeKind)
-            => TypeResolver.GetTypeName(r, handle) ?? "object";
-        public string GetTypeFromSerializedName(string name)
-        {
-            int comma = name.IndexOf(',');
-            return comma >= 0 ? name[..comma] : name;
-        }
-
-        public PrimitiveTypeCode GetUnderlyingEnumType(string type)
-        {
-            foreach (var handle in reader.TypeDefinitions)
-            {
-                var def = reader.GetTypeDefinition(handle);
-                if (TypeResolver.GetFullName(reader, def) != type)
-                    continue;
-                foreach (var fieldHandle in def.GetFields())
-                {
-                    var field = reader.GetFieldDefinition(fieldHandle);
-                    if ((field.Attributes & FieldAttributes.Static) == 0)
-                        return field.DecodeSignature(new PrimitiveCodeProvider(), null);
-                }
-            }
-            return PrimitiveTypeCode.Int32;
-        }
-    }
-
-    /// <summary>Minimal signature provider that reports an enum's underlying primitive type code.</summary>
-    sealed class PrimitiveCodeProvider : ISignatureTypeProvider<PrimitiveTypeCode, object?>
-    {
-        public PrimitiveTypeCode GetPrimitiveType(PrimitiveTypeCode code) => code;
-        public PrimitiveTypeCode GetTypeFromDefinition(MetadataReader r, TypeDefinitionHandle h, byte k) => PrimitiveTypeCode.Int32;
-        public PrimitiveTypeCode GetTypeFromReference(MetadataReader r, TypeReferenceHandle h, byte k) => PrimitiveTypeCode.Int32;
-        public PrimitiveTypeCode GetSZArrayType(PrimitiveTypeCode e) => PrimitiveTypeCode.Int32;
-        public PrimitiveTypeCode GetArrayType(PrimitiveTypeCode e, ArrayShape s) => PrimitiveTypeCode.Int32;
-        public PrimitiveTypeCode GetByReferenceType(PrimitiveTypeCode e) => PrimitiveTypeCode.Int32;
-        public PrimitiveTypeCode GetPointerType(PrimitiveTypeCode e) => PrimitiveTypeCode.Int32;
-        public PrimitiveTypeCode GetGenericInstantiation(PrimitiveTypeCode g, ImmutableArray<PrimitiveTypeCode> a) => PrimitiveTypeCode.Int32;
-        public PrimitiveTypeCode GetGenericMethodParameter(object? ctx, int i) => PrimitiveTypeCode.Int32;
-        public PrimitiveTypeCode GetGenericTypeParameter(object? ctx, int i) => PrimitiveTypeCode.Int32;
-        public PrimitiveTypeCode GetModifiedType(PrimitiveTypeCode m, PrimitiveTypeCode u, bool r) => u;
-        public PrimitiveTypeCode GetPinnedType(PrimitiveTypeCode e) => e;
-        public PrimitiveTypeCode GetFunctionPointerType(MethodSignature<PrimitiveTypeCode> s) => PrimitiveTypeCode.Int32;
-        public PrimitiveTypeCode GetTypeFromSpecification(MetadataReader r, object? ctx, TypeSpecificationHandle h, byte k) => PrimitiveTypeCode.Int32;
-    }
 
     private static string? TryGetAttributeDisplayValue(MetadataReader reader, CustomAttribute attr)
     {
